Move `onAttachmentIsEditableChange`
